Add `LocationSettingsScreen` and move corresponding features from `CredentialSettingsScreen`
